+++ lib/classes/class-settings.php
@@ -46,7 +46,8 @@ namespace UsabilityDynamics\WPP {
           'thumbnail_size' => 'tiny_thumb',
           'fancybox_preview' => 'true',
           'display_slideshow' => 'false',
-          'show_children' => 'true'
+          'show_children' => 'true',
+          'pagination_type' => 'numeric'
         );
 
         $data['configuration']['single_property_view'] = array(